+++ js/components/arrow.js
@@ -1,5 +1,5 @@
-import {render} from '../../utils/render';
-import store from '../../store';
+import store from '../store';
+import {render} from '../utils/render';
 
 const listeners = [
   {
@@ -11,7 +11,7 @@ const listeners = [
   }
 ];
 
-const arrow = () => {
+const Arrow = () => {
   const content = `
     <span class="visually-hidden">Вернуться к началу</span>
   <svg class="icon" width="45" height="45" viewBox="0 0 45 45" fill="#000000">
@@ -31,4 +31,4 @@ const arrow = () => {
   });
 };
 
-export default arrow;
+export default Arrow;
