+++ js/components/game-3.js
@@ -1,32 +1,30 @@
-// import {createElementFromTemplate, validate} from '../utils/main';
-import {render} from '../../utils/render';
+// import {createElementFromTemplate} from '../utils/main';
+import {render} from '../utils/render';
 // import {pubSub} from '../pubSub';
-import store from '../../store';
+import store from '../store';
 
 const listeners = [
   {
     targetSelector: `.game__content`,
     type: `change`,
     callback: () => {
-      // return validate(form, [`question1`]) && pubSub.publish(`changeScreen`, `game3`);
+      // pubSub.publish(`changeScreen`, `stats`);
     }
   }
 ];
 
-const Game2 = ({answers}) => {
+const Game3 = ({answers}) => {
   const content = `
-    <p class="game__task">Угадай, фото или рисунок?</p>
-    <form class="game__content  game__content--wide">
+    <p class="game__task">Найдите рисунок среди изображений</p>
+    <form class="game__content  game__content--triple">
       <div class="game__option">
-        <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
-        <label class="game__answer  game__answer--photo">
-          <input class="visually-hidden" name="question1" type="radio" value="photo">
-          <span>Фото</span>
-        </label>
-        <label class="game__answer  game__answer--paint">
-          <input class="visually-hidden" name="question1" type="radio" value="paint">
-          <span>Рисунок</span>
-        </label>
+        <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
+      </div>
+      <div class="game__option  game__option--selected">
+        <img src="http://placehold.it/304x455" alt="Option 2" width="304" height="455">
+      </div>
+      <div class="game__option">
+        <img src="http://placehold.it/304x455" alt="Option 3" width="304" height="455">
       </div>
     </form>
     <ul class="stats">
@@ -46,11 +44,11 @@ const Game2 = ({answers}) => {
 
   render({
     node: `section`,
-    id: `game2`,
+    id: `game3`,
     className: `game`,
     template: content,
     listeners
   });
 };
 
-export default store.connect(Game2, [`answers`]);
+export default store.connect(Game3, [`answers`]);
